+++ mocap_popy/aux_scripts/generate_template_file_mapping.py
